remove most -Wsign-compare warnings

+++ libclamav/7z/CpuArch.h
@@ -18,7 +18,7 @@ If MY_CPU_LE_UNALIGN is not defined, we don't know about these properties of pla
 */
 
 #define GetUi16(p) (cli_readint16(p))
-#define GetUi32(p) (cli_readint32(p))
+#define GetUi32(p) ((UInt32)cli_readint32(p))
 
 #define GetUi64(p) (GetUi32(p) | ((UInt64)GetUi32(((const Byte *)(p)) + 4) << 32))
